requirements.txt
firebase-admin
schedule
openai
python-dotenv